requirements.txt
python-docx==0.8.11
pandas==2.0.1
PyPDF2==3.0.1
xlrd==2.0.1
openpyxl==3.1.2
xls2xlsx==0.2.0
beautifulsoup4==4.12.2
